+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.16.1] - 2025-11-17
+
+### Fixed
+
+- enable scrolling for teleport dialog to fix inaccessible Create button (Issue #34)
+
+
 ## [1.16.0] - 2025-11-17
 
 ### Added